# Utils Added new python subpackage to do some simple tasks users may regularly need
# Models
All models now return a list output intead of tuples. This is to make it easier to deal with failed models since user can check result[0]>0 or len(result)>0 to see if model succeeded whereas previous result would not be a list on a fail and would on a success which is awkward.

# Docs
- Minor typo fixes and updated all notebooks to make sure they match latest version of the code.

- Not technically a doc but updated examples to new code